+++ NEWS.md
@@ -1,3 +1,18 @@
+# rjsoncons 1.2.0
+
+- (1.1.0.9500) update documentation, include NDJSON-specific, web-only
+  vignette
+- (1.1.0.9400) support NDJSON and file / url connections
+- (1.1.0.9300) implement `j_query()` (query without requiring path
+  specification), `j_pivot()`, and `j_path_type()`. Remove
+  `jsonpivot()`.
+- (1.1.0.9200) implement `jsonpointer()` for querying JSON documents.
+- (1.1.0.9100) update jsoncons library to 173.2, relaxing compiler
+  requirements to c++11.
+- (1.1.0.9000) implement `jsonpivot()` to transform JSON
+  array-of-objects to object-of-arrays, a common step before
+  representation as a data.frame.
+
 # rjsoncons 1.1.0
 
 - (1.1.0) CRAN release
@@ -19,7 +34,7 @@
 - (0.0.3) support object names ordering 'asis' or 'sort'
 - (0.0.3) DESCRIPTION file updates: correct 'Title:' capitalization;
   avoid warnings about most misspellings
-- (0.0.3) Add github action to rebuild README.md from
+- (0.0.3) Add GitHub action to rebuild README.md from
   vignettes/rjsoncons.Rmd
 - (0.0.2) jsoncons library update
 - (0.0.2) support for R object query in addition to JSON string